Move EC_PhysicsSystem and EC_LuaScriptingSystem into their sibling folders
Engine/Subsystems/ had already-organized CollisionSystems/ and
Scripting/ subdirectories sitting alongside loose top-level files -
including these two, which each belong in one of those two folders
rather than at the Subsystems root: EC_PhysicsSystem is the resolution/
dynamics counterpart to CollisionSystems/EC_PhysicsResolution.h, and
EC_LuaScriptingSystem orchestrates the Scripting/*API classes.

EC_CameraSystem, EC_SpatialSystem, EC_TransformSystem, ControlSystem,
and the shared EC_System/EC_SystemType base stay at Subsystems root -
each is a standalone system with no closer relative to group with.

First of two stages agreed with the user for cleaning up directory
structure; Graphics/ (52 flat files) is next, larger and separate.
Per-directory CMakeLists.txt files replacing the top-level
GLOB_RECURSE are an explicit follow-up once both directory stages
land, not part of either.
