+++ src/Engine/EC_Engine.cpp
@@ -2,9 +2,9 @@
 #include "Engine/Subsystems/EC_SpatialSystem.h"
 #include "Engine/Subsystems/EC_TransformSystem.h"
 #include "Engine/Subsystems/EC_CameraSystem.h"
-#include "Engine/Subsystems/EC_LuaScriptingSystem.h"
+#include "Engine/Subsystems/Scripting/EC_LuaScriptingSystem.h"
 #include "Engine/Subsystems/CollisionSystems/EC_CollisionSystem.h"
-#include "Engine/Subsystems/EC_PhysicsSystem.h"
+#include "Engine/Subsystems/CollisionSystems/EC_PhysicsSystem.h"
 #include "TaskManager/EC_PhysicsThreadTask.h"
 #include "TaskManager/EC_ScriptingTask.h"
 #include "xml/XML.h"
